% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_screening.R
\name{spearman_screen}
\alias{spearman_screen}
\title{Estimating equation for ELCIC under GLM}
\usage{
spearman_screen(index_data,data_vessel,size)
}
\arguments{
\item{index_data}{A matrix containing covariates. The first column should be all ones corresponding to the intercept. See more details in}

\item{data_vessel}{A vector containing outcomes.}

\item{size}{A plug-in estimator solved by an external estimating procedure.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equation for ELCIC under GLM. This estimating equation is used for marginal mean selection.
}
