% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_screening.R
\name{test_screen}
\alias{test_screen}
\title{Estimating equation for ELCIC under GLM}
\usage{
test_screen(data_vessel,group,covariate)
}
\arguments{
\item{data_vessel}{A matrix containing covariates. The first column should be all ones corresponding to the intercept. See more details in}

\item{group}{A vector containing outcomes.}

\item{covariate}{A plug-in estimator solved by an external estimating procedure.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equation for ELCIC under GLM. This estimating equation is used for marginal mean selection.
}
