% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_learning.R
\name{vc.fit}
\alias{vc.fit}
\title{Fit a varying coefficient model}
\usage{
vc.fit(agent,data_observe,x_cov)
}
\arguments{
\item{agent}{The imputed disease risk.}

\item{data_observe}{The gene expression matrix.}

\item{x_cov}{The vector including covariate values (eg., smoking: 0 and 1).}
}
\value{
A list of two matrices including varying intercept and varying covariate effect, respective.
}
\description{
This function fit the gene expression data into a varying coefficient model.
}
