% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_learning.R
\name{base.construct}
\alias{base.construct}
\title{Estimating equation for ELCIC under GLM}
\usage{
base.construct(data_observe,data_fitted, degree=3,
len.knots=3,data_fitted_cov,agent,x_cov)
}
\arguments{
\item{data_observe}{A matrix containing covariates. The first column should be all ones corresponding to the intercept. See more details in}

\item{data_fitted}{A vector containing outcomes.}

\item{degree}{A plug-in estimator solved by an external estimating procedure.}

\item{len.knots}{A plug-in estimator solved by an external estimating procedure.}

\item{data_fitted_cov}{A plug-in estimator solved by an external estimating procedure.}

\item{agent}{A plug-in estimator solved by an external estimating procedure.}

\item{x_cov}{A plug-in estimator solved by an external estimating procedure.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equation for ELCIC under GLM. This estimating equation is used for marginal mean selection.
}
