% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_learning.R
\name{network.learn}
\alias{network.learn}
\title{Estimating equation for ELCIC under GLM}
\usage{
network.learn(data_observe,x_cov,X_big_int,X_big_int_cov,
agent,degree=3,len.knots=3,cv=TRUE,nfolds=20,alpha=1)
}
\arguments{
\item{data_observe}{A matrix containing covariates. The first column should be all ones corresponding to the intercept. See more details in}

\item{x_cov}{A plug-in estimator solved by an external estimating procedure.}

\item{X_big_int}{A vector containing outcomes.}

\item{X_big_int_cov}{A vector containing outcomes.}

\item{agent}{A plug-in estimator solved by an external estimating procedure.}

\item{degree}{A plug-in estimator solved by an external estimating procedure.}

\item{len.knots}{A plug-in estimator solved by an external estimating procedure.}

\item{cv}{A plug-in estimator solved by an external estimating procedure.}

\item{nfolds}{A plug-in estimator solved by an external estimating procedure.}

\item{alpha}{A plug-in estimator solved by an external estimating procedure.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equation for ELCIC under GLM. This estimating equation is used for marginal mean selection.
}
