% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_screening.R
\name{spearman_screen}
\alias{spearman_screen}
\title{Gene screening based on the Spearman correlation.}
\usage{
spearman_screen(agent,data_vessel,size)
}
\arguments{
\item{agent}{The imputed disease risk.}

\item{data_vessel}{The gene expression matrix.}

\item{size}{The number of selected genes.}
}
\value{
The vector containing the index for the selected genes.
}
\description{
This function can be used to screen the genes based on Spearman correlation coefficient.
}
