% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_screening.R
\name{test_screen}
\alias{test_screen}
\title{Gene selection based on the test of differential gene expression.}
\usage{
test_screen(data_vessel,group,covariate)
}
\arguments{
\item{data_vessel}{The gene expression matrix.}

\item{group}{The vector including disease status (0 and 1).}

\item{covariate}{The vector including covariate values (eg., smoking: 0 and 1).}
}
\value{
The vector containing the index for the selected genes.
}
\description{
This function can be used to select the genes based on the test of differential gene expression.
}
