% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_learning.R
\name{vc.fit}
\alias{vc.fit}
\title{Estimating equation for ELCIC under GLM}
\usage{
vc.fit(agent,data_observe,x_cov)
}
\arguments{
\item{agent}{A plug-in estimator solved by an external estimating procedure.}

\item{data_observe}{A matrix containing covariates. The first column should be all ones corresponding to the intercept. See more details in}

\item{x_cov}{A plug-in estimator solved by an external estimating procedure.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equation for ELCIC under GLM. This estimating equation is used for marginal mean selection.
}
